Add ability to go to previous template
Updated the code to now respond to shift+tab, which is interpreted as
"KEY_BTAB" in curses, to move to the previous template.
Changed the logic of the code to remove the index of the template
we're editing from the position list and store it in it's own variable.
Depending on the direction of travel we add the current index to one
end of the index list and take from the other.
A positive integer  indicates we're moving down the snippet to the next
template after this current one.
A negative integer indicates we're moving up the snippet to the previous
template before this one.