Add argparser code and update variables to constants
Started the code to make the tool more dynamic. It now requires a file
containing a snippet to be passed to it. Right now the snippet is
ignored but this will be updated in a later commit.
Also updated variables in the global scope to be constants to stop
pylint complaining.